+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # Changelog  
+## Version 0.68.1 - Unreleased
+🐞Fix grouped shape X, Y, Width and Height [#599](https://github.com/ShapeCrawler/ShapeCrawler/issues/599)  
 
 ## Version 0.68.0 - 2025-05-08
 🍀Added support adding Scatter Chart [#65](https://github.com/ShapeCrawler/ShapeCrawler/issues/65)